Fix for defaults.write
The key_type dictionary was iterated over with no order. This causes a
boolean and possibly other values to be identified as other types such as `int`
should python match it first within the `isinstance()` return check.

To correct this, the `sorted()` function was added to respect the order.

This also should fix #34.